fix: set latest docker tag for only tagged commits where tag starts with 'v'
In our current CI flow, we add 'latest' docker tag for the images we push to the master or main branches, but master branch images are deployed to dev environment, not prod. Ideally, we should mark images deployed to prod environment with latest tag, and mark images that are deployed to the dev env. with the branch name

see: BE-410